move drawer and fix search lag
